Rename cards/printer → cards/parallel; README/CLAUDE.md printer docs
- Rename src/core/cards/printer/ to src/core/cards/parallel/ — the directory
  name was misleading since SSC also drives printers (IW-I/II); the parallel
  card is a Centronics interface, not a printer-specific card
- Update include path in emulator.hpp and both CMakeLists.txt entries
- Drop synthetic ROM fallback from parallel_card — real 341-0057 PROM is
  always present; buildROM() is now a single memcpy
- Fix Centronics parallel port section header (was "Printer interface")
- README: add printer to features, parallel card and SSC to slot table,
  parallel card description, 341-0057 ROM entry, remove parallel card from TODO
- CLAUDE.md: add cards/parallel/ and cards/ssc printer notes, 341-0057 ROM,
  parallel card to Available Cards list and directory tree
- slot-configuration-window: align JS fallback defaults with WASM defaults
  (mockingboard/smartport/disk2/thunderclock); drop stale parallel from fallback
