+++ src/core/cards/parallel/parallel_card.cpp
@@ -17,40 +17,14 @@ ParallelCard::ParallelCard() {
 }
 
 void ParallelCard::buildROM() {
-    // Authentic firmware. The real Apple Parallel Interface Card PROM (341-0057)
-    // contains TWO 256-byte firmwares; on the real card DIP SW1:6 selects which
-    // one maps to $Cn00 — "Parallel Printer" (341-0005, auto-LF after CR) in the
-    // upper half at PROM offset $100, "Centronics" (341-0019, no auto-LF) in the
-    // lower half. We default to Parallel Printer (no DIP yet). Apple peripheral
-    // firmware self-locates its slot via $C080,X indexing, so the ROM image is
-    // slot-independent — no per-slot patching.
+    // The real Apple Parallel Interface Card PROM (341-0057) contains two
+    // 256-byte firmwares; DIP SW1:6 selects which maps to $Cn00 —
+    // "Parallel Printer" (341-0005, auto-LF after CR) in the upper half at
+    // PROM offset $100, "Centronics" (341-0019, no auto-LF) in the lower half.
+    // We default to Parallel Printer. Apple peripheral firmware self-locates
+    // its slot via $C080,X indexing, so the ROM image is slot-independent.
     constexpr size_t kPrinterBase = 0x100; // 341-0005 "Parallel Printer" half
-    if (roms::ROM_PARALLEL_SIZE >= kPrinterBase + sizeof(rom_)) {
-        memcpy(rom_, roms::ROM_PARALLEL + kPrinterBase, sizeof(rom_));
-        usingRealRom_ = true;
-        return;
-    }
-
-    // Fallback: synthetic per-character output handler, used only when the ROM
-    // file is absent from the build.
-    //
-    // Apple II output convention: PR#n sets CSW ($36/$37) = $Cn00, then COUT
-    // does JMP ($36) into here ONCE PER CHARACTER with the char in A (high bit
-    // set). We emit it via a bare STA to the data port (which fires the print
-    // callback) and RTS. STA preserves A/X/Y, so the protocol contract that the
-    // character survives the call is honoured. The JS printer strips the high
-    // bit. No two-stage revectoring — that swallowed the first character and
-    // clobbered A.
-    usingRealRom_ = false;
-    memset(rom_, 0xFF, sizeof(rom_));
-
-    // I/O base: slot 1 = $C090, slot 2 = $C0A0, etc.
-    uint16_t dataPort = 0xC080 + static_cast<uint16_t>(slotNumber_) * 0x10;
-    int i = 0x00;
-    rom_[i++] = 0x8D;                                          // STA dataPort (abs)
-    rom_[i++] = static_cast<uint8_t>(dataPort & 0xFF);
-    rom_[i++] = static_cast<uint8_t>(dataPort >> 8);
-    rom_[i++] = 0x60;                                          // RTS
+    memcpy(rom_, roms::ROM_PARALLEL + kPrinterBase, sizeof(rom_));
 }
 
 uint8_t ParallelCard::readIO(uint8_t offset) {
