+++ src/js/ui/slot-configuration-window.js
@@ -553,7 +553,7 @@ export class SlotConfigurationWindow extends BaseWindow {
 
   async applyInitialSettings() {
     const saved = this.loadSettingsFromStorage();
-    const config = saved || { 1: "parallel", 5: "smartport", 7: "thunderclock" };
+    const config = saved || { 4: "mockingboard", 5: "smartport", 6: "disk2", 7: "thunderclock" };
     if (this.wasmModule && this.wasmModule._setSlotCard) {
       for (const [slot, cardId] of Object.entries(config)) {
         const slotNum = parseInt(slot, 10);
